chore: change `rust-toolchain` file to toml format
The `rust-toolchain` file containing only a Rust version number is
deprecated in favor of a TOML-formatted `rust-toolchain.toml`. Using the
old format seems to make Dependabot unhappy --- it complains that:

```
only rust-toolchain files formatted as TOML are supported, the non-TOML
format was deprecated by Rust
```

Therefore, this branch changes the toolchain file in this repo to the
TOML format. This required updating the CI workflows that check that
the toolchain matches to use a new regex.