ci: fix default Rust toolchain in `test-policy`
PR #11528 changed the `rust-toolchain` file to use the TOML format
rather than the plaintext format. Unfortunately, this broke the
`test-policy` CI job, which `cat`s the contents of the `rust-toolchain`
file to determine which Rust version to install, and that file now no
longer exists (since it's now in `rust-toolchain.toml`). See:
https://github.com/linkerd/linkerd2/actions/runs/6649430068/job/18068192854?pr=11535

This branch fixes that by changing the CI job to use a regex to extract
the Rust toolchain version instead.