Fix GetProfiles error when address is opaque and unmeshed (#11556)
When we do a GetProfile lookup for an opaque port on an unmeshed pod,
we attempt to look up the inbound listen port of that pod's proxy. Since
that pod has no proxy, this fails and we return an error to the GetProfile
API call. This causes the proxy to fail to be able to resolve the profile and
be unable to route the traffic.

We revert to the previous behavior of only logging when we cannot look
up the inbound listen port instead of returning an error.

Signed-off-by: Alex Leong <alex@buoyant.io>