service-mirror: support gateway resolving to multiple addresses (#11499)
If the gateway address was a name resolving to multiple addresses,
previously service-mirror took only one of those address. This behavior
could led to downtime in case the gateway behind the IP picked by service-mirror
is down. This commit fix this behavior by considering all the IPs
resolved. 

Now that we consider all the IPs resolved by the DNS, we make sure they
are sorted lexicographically so that resolveGatewayAddress answer a
stable output.

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@ledger.fr>