Change behaviour gateway addressess override in Link command (#11564)
When linking two clusters, a `--gateway-addresses` override can be
passed to the command. If applied, the override populates the Link
resource with a set of addresses received from the flag, instead of
reading them from Linkerd's Gateway LoadBalancer service. Currently, the
override is used only when the LoadBalancer service does not exist (or
does not contain any IPs).

An override should take precedence over default behaviour. As it stands,
the user experience is confusing; it can also make manual testing for
changes hard (when an override has to be passed) and it can inhibit
users from passing in a static IP or DNS addresses to avoid relying on
dynamic IPs in the LoadBalancer service.

This change makes the override behave as expected.

Signed-off-by: Matei David <matei@buoyant.io>