policy: allow spiffe ids in `MeshTLSAuthentication` (#11882)
This change enables the use of SPIFFE identities in `MeshTLSAuthentication`.
To make that happen validation of the identities field on the CRD has been moved
to the policy controller admission webhook. Apart from a more clear expression
of the constraints that a SPIFFE ID needs to meet, this approach allows for
richer error messages. Note that the DNS validation is still based on a regex.

Signed-off-by: Zahari Dichev <zaharidichev@gmail.com>