Improve server-side validation for ExternalWorkload
We introduced an ExternalWorkload CRD along with bindings for mesh
expansion. Currently, the CRD allows users to create ExternalWorkload
resources without adding a meshTls strategy.

This change adds some more validation restrictions to the CRD definition
(i.e. server side validation). When a meshTls strategy is used, we
require both identity and serverName to be present. Furthermore, we
require that when a list of ports is specified, each item in the list
must contain a numeric port.

The change also touches on the bindings (Go & Rust types) to remove
optionality from meshTls fields.

Signed-off-by: Matei David <matei@buoyant.io>