Bump CNI plugin version to v1.3.0 in Go pkg (#11912)
We released a new version of the CNI plugin. The chart has been updated
to reference the new version, however, some of the tests and the Go
`version` pkg still reference the old version (v1.2.2). When installing
through the CLI, I noticed that even though the chart value renders an
image for the new repair controller, the image used is still v1.2.2, and
as such, the container won't be started due to a missing binary.

This change bumps the version to v1.3.0 everywhere.

Signed-off-by: Matei David <matei@buoyant.io>