Index ExternalWorkload resources in the policy controller
ExternalWorkload resources represent as a resource configuration
associated with a process (or a group of processes) that are foreign to
a Kubernetes cluster. It allows Linkerd to read / write and store
configuration for mesh expansion. Since VMs will be able to receive
inbound traffic from a variety of resources, the proxy should be able to
dynamically discover inbound authorisation policies.

This change introduces a set of callbacks in the indexer that will apply
(or delete) ExternalWorkload resources. In addition, we ensure that
ExternalWorkloads can be processed in a similar fashion to pods (where
applicable, of course) wrt to server matching and defaulting.

To serve discovery requests for a VM, the policy controller will need
to:

* Start a background task that will run the watchers & associated
  callbacks
* Perform discovery when requests reference an external workload

(both changes will come as follow-ups)

A quick list of changes:

* ExternalWorkloads can now be indexed in the inbound (policy) index.
* Renamed the pod module in the inbound index to be more generic
  ("workload"); the module has some re-usable building blocks that we
  can use for external workloads.
* Moved common functions (e.g. building a default inbound server) around
  to share what's already been done without abstracting more or
  introducing generics.

Signed-off-by: Matei David <matei@buoyant.io>