Relax validation for ExternalWorkload Status fields (#11979)
ExternalWorkload resources require that status condition has almost all of its
fields set (with the exception of a date field). The original inspiration for
this design was the HTTPRoute object.

When using the resource, it is more practical to handle many of the fields as
optional; it is cumbersome to fill out the fields when creating an
ExternalWorkload. We change the settings to be in-line with a [Pod] object
instead.

[Pod]:
https://github.com/kubernetes/api/blob/7d1a2f7a7336547258f7dc8594d7e35f0399e966/core/v1/types.go#L3063-L3084


---------

Signed-off-by: Matei David <matei@buoyant.io>