Restrict the number of IPs an ExternalWorkload can have
A Kubernetes pod may be assigned at [most one IP address][pod-docs]
for each supported protocol (i.e. IPv6 and IPv4), without the use of
specialised CNIs or network configurations. When processing
addresses in an endpoint, we will only ever use one address.

ExternalWorkload resources have a generic workloadIPs field that
allow any number of addresses to be added. We want the behaviour to
be similar to a pod -- only one address (of each protcol) should be
used for routing.

We restrict the CRD server-side validation to allow only one IP
address. Since we do not yet support IPv6, this will ensure that two
IPv4 addresses will not be declared by the same workload. Once IPv6
support lands, or once we have a dedicated validator, we can relax
the CRD validation.

[pod-docs]:  https://pkg.go.dev/k8s.io/kubernetes@v1.29.1/pkg/apis/core#PodStatus

Signed-off-by: Matei David <matei@buoyant.io>