Limit service publisher updates from server by namespace (#12040)
A Server may only select workloads in its own namespace. Therefore, when the
destination controller receives an update for a Server, it only needs to
potentially send updates to watches on workloads in that same namespace. Taking
this into account allows us avoid all opaqueness computations for workloads in
other namespaces.

Signed-off-by: Alex Leong <alex@buoyant.io>