build: Support alternate proxy artifact sources
It's necessary in some testing environments to use a proxy artifact
from a different source than the default.

This change updates the proxy fetching process in a few ways:

1. The `LINKERD2_PROXY_REPO` environment variable can be set to override the
   default proxy artifact source. The default is 'linkerd/linkerd2-proxy'.
2. The `LINKERD2_PROXY_VERSION` environment variable can be set to override the
   default proxy artifact version. The build script sets this value from the
   .proxy-version file.
3. The `GITHUB_TOKEN` environment variable can be set to provide a GitHub
   personal access token for fetching the proxy artifact. This is useful when
   the proxy artifact is hosted in a private repository.

To support this, the fetch-proxy script has been updated to use the GitHub API,
processed with jq. The Dockerfile-proxy has been updated to include jq and curl
in its fetch process.