build: Update fetch-proxy to support alternate repos
In some build and test scenarios, it's desirable to fetch alternate proxy
releases during the build process.

This change updates the proxy container image build tooling to support the
LINKERD2_PROXY_REPO and LINKERD2_PROXY_GITHUB_TOKEN environment variables. These
may be set to the desired repository and a GitHub personal access token,
respectively.

When these are unset, the default behavior is unchanged.