Add support for loadBalancerCluster to multi-cluster gateway service
The multicluster helm chart currently does not support the `loadBalancerClass` field. This is required by some loadbalance services like Tailscale.

Add a conditional `loadBalancerClass` key to the helm chart values and gateway svc.

Once merged, deploy the multicluster helm chart and configure `.Values.gateway.loadBalancerClass` to a value, ensure that is part of the resulting resource. Unset the value, ensure it is _not_ part of the resulting resource.

Fixes #12100

Signed-off-by: Adam Toy <atoy3731@gmail.com>