ci: Update release workflow to generate release notes
GitHub can automatically generate release notes from commit messages. This saves
us from having to manually update the `CHANGES.md` file every week.

In preparation for the next edge release, this change updates the release
workflow to stop relying on the CHANGES file. The CHANGES file is updated with a
pointer to the releases page.