Update version parsing to work with build info
Problem

The version checking that's used when checking if Linkerd is running the
latest version treats strings with the same version but different build
info as unqeual, when they should be considered equal. For instance, the
versions `edge-23.1.3` and `edge-23.1.3-arm` should be treated as equal.

Solution

Update channel version parsing to ignore build info in versions,
unless the build info is a numeric hotpatch, in which case the hotpatch
is preserved and the channel is updated to reflect that.

Validation

See the updated tests in the `version` package.

Signed-off-by: Kevin Ingelman <ki@buoyant.io>