ci: Ensure integration status is set
We have been using a path filter for PRs to avoid running integration tests
needlessly. However, this prevents us from having a status check that can be
satisfied (i.e. for the purposes of having required checks).

This commit updates the integration workflow to match paths within the workflow,
and it adds a final *integrations-ok* job that succeeds as long as no
integrations failed.